+++ include/cch.hpp
@@ -303,7 +303,7 @@ private:
             // may stay in registers across chunks.  src is a std::byte span,
             // and std::byte is exempt from strict aliasing, so the compiler
             // cannot otherwise rule out state_ and src overlapping.
-            if (std::size(src) >= get_state_size_bytes())
+            if (std::size(src) >= static_cast<std::size_t>(get_state_size_bytes()))
             {
                 state_t state = state_;
                 auto absorbs_since_mix = absorbs_since_mix_;
@@ -320,7 +320,7 @@ private:
                     }
 
                     src = src.subspan(get_state_size_bytes());
-                } while (std::size(src) >= get_state_size_bytes());
+                } while (std::size(src) >= static_cast<std::size_t>(get_state_size_bytes()));
 
                 state_ = state;
                 absorbs_since_mix_ = absorbs_since_mix;
@@ -330,7 +330,7 @@ private:
         {
             // Then, process whole chunks directly from the source, bypassing the
             // input buffer.
-            while (std::size(src) >= get_state_size_bytes())
+            while (std::size(src) >= static_cast<std::size_t>(get_state_size_bytes()))
             {
                 absorb_(src);
